+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 ### Added
 - Whitelist functionality for discover page
 - Unified solution for preprint words
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Styling and format of the branded navbar to match current styling in ember-osf
 - Reduced number of calls to the `preprints_provider` endpoint
+- Sharing from the preprint detail view to use ember-osf sharing icons
 
 ## [0.118.5] - 2018-04-27
 ### Changed